% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{ox_view_pptx}
\alias{ox_view_pptx}
\title{Preview a .pptx file}
\usage{
ox_view_pptx(x, interactive = NA, browser = FALSE)
}
\arguments{
\item{x}{Path to an existing .pptx file.}

\item{interactive}{Whether to actually open the viewer. Defaults to
\code{interactive()}; set explicitly to avoid depending on session state.}

\item{browser}{If \code{TRUE}, force opening in the system browser instead
of the RStudio Viewer pane.}
}
\description{
Preview a .pptx file
}
